Move around type traits for buffer, scratch, and ptr
