+++ ttg/ttg/device/task.h
@@ -18,7 +18,8 @@ namespace ttg::device {
 
     struct device_input_data_t {
       using impl_data_t = decltype(TTG_IMPL_NS::buffer_data(std::declval<ttg::Buffer<int>>()));
-
+      : impl_data(data), scope(scope), is_const(is_const), is_scratch(is_scratch)
+      { }
       impl_data_t impl_data;
       ttg::scope scope;
       bool is_const;
@@ -33,11 +34,12 @@ namespace ttg::device {
     /* extract buffer information from to_device_t */
     template<typename... Ts, std::size_t... Is>
     auto extract_buffer_data(detail::to_device_t<Ts...>& a, std::index_sequence<Is...>) {
+      using arg_types = std::tuple<Ts...>;
       return std::array{
                 device_input_data_t{TTG_IMPL_NS::buffer_data(std::get<Is>(a.ties)),
                                     std::get<Is>(a.ties).scope(),
-                                    std::is_const_v<std::tuple_element<Is, decltype(a.ties)>>,
-                                    ttg::meta::is_devicescratch_v<std::tuple_element<Is, decltype(a.ties)>>}...};
+                                    ttg::meta::is_const_v<std::tuple_element<Is, arg_types>>,
+                                    ttg::meta::is_devicescratch_v<std::tuple_element<Is, arg_types>>}...};
     }
   }  // namespace detail
