+++ ttg/ttg/devicescratch.h
@@ -3,6 +3,7 @@
 
 #include "ttg/devicescope.h"
 #include "ttg/fwd.h"
+#include "ttg/util/meta.h"
 
 namespace ttg {
 
@@ -13,6 +14,24 @@ template<typename T>
 auto make_scratch(T* val, ttg::scope scope, std::size_t count = 1) {
   return devicescratch<T>(val, scope, count);
 }
+
+namespace meta {
+
+  /* Specialize some traits */
+
+  template<typename T>
+  struct is_devicescratch<ttg::devicescratch<T>> : std::true_type
+  { };
+
+  template<typename T>
+  struct is_devicescratch<const ttg::devicescratch<T>> : std::true_type
+  { };
+
+  template<typename T>
+  struct is_const<ttg::devicescratch<T>> : std::is_const<T>
+  { };
+
+} // namespace meta
 
 } // namespace ttg
